Extract PaymentIntentActivity super class
