+++ example/src/main/java/com/stripe/example/activity/PaymentIntentActivity.kt
@@ -1,0 +1,94 @@
+package com.stripe.example.activity
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.ViewModelProvider
+import com.stripe.android.ApiResultCallback
+import com.stripe.android.PaymentConfiguration
+import com.stripe.android.PaymentIntentResult
+import com.stripe.android.Stripe
+import com.stripe.android.model.ConfirmPaymentIntentParams
+import com.stripe.android.model.PaymentMethodCreateParams
+import com.stripe.example.R
+import com.stripe.example.module.PaymentIntentViewModel
+import org.json.JSONObject
+import java.lang.ref.WeakReference
+
+/**
+ * Base class for Activity's that wish to create and confirm payment methods.
+ * Subclasses should observer on the [PaymentIntentViewModel]'s LiveData properties
+ * in order to display state on the interaction.
+ */
+abstract class PaymentIntentActivity : AppCompatActivity() {
+    internal val viewModel: PaymentIntentViewModel by lazy {
+        ViewModelProvider(this,
+            ViewModelProvider.AndroidViewModelFactory(application)
+        )[PaymentIntentViewModel::class.java]
+    }
+    protected val stripe: Stripe by lazy {
+        Stripe(this, PaymentConfiguration.getInstance(this).publishableKey)
+    }
+    private val keyboardController: KeyboardController by lazy {
+        KeyboardController(this)
+    }
+
+    protected fun createAndConfirmPaymentIntent(
+        country: String,
+        paymentMethodCreateParams: PaymentMethodCreateParams
+    ) {
+        keyboardController.hide()
+
+        viewModel.createPaymentIntent(country) {
+            handleCreatePaymentIntentResponse(it, paymentMethodCreateParams)
+        }
+    }
+
+    private fun handleCreatePaymentIntentResponse(
+        responseData: JSONObject,
+        params: PaymentMethodCreateParams
+    ) {
+        val secret = responseData.getString("secret")
+        viewModel.status.value += "\n\nStarting PaymentIntent confirmation"
+        stripe.confirmPayment(
+            this,
+            ConfirmPaymentIntentParams.createWithPaymentMethodCreateParams(
+                paymentMethodCreateParams = params,
+                clientSecret = secret,
+                returnUrl = "example://return_url"
+            )
+        )
+    }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        stripe.onPaymentResult(requestCode, data, PaymentIntentResultCallback(this))
+    }
+
+    protected fun onConfirmSuccess(result: PaymentIntentResult) {
+        val paymentIntent = result.intent
+        viewModel.status.value += "\n\n" +
+            "PaymentIntent confirmation outcome: ${result.outcome}\n\n" +
+            getString(R.string.payment_intent_status, paymentIntent.status)
+        viewModel.inProgress.value = false
+    }
+
+    protected fun onConfirmError(e: Exception) {
+        viewModel.status.value += "\n\nException: " + e.message
+        viewModel.inProgress.value = false
+    }
+
+    internal class PaymentIntentResultCallback(
+        activity: PaymentIntentActivity
+    ) : ApiResultCallback<PaymentIntentResult> {
+
+        private val activityRef = WeakReference(activity)
+
+        override fun onSuccess(result: PaymentIntentResult) {
+            activityRef.get()?.onConfirmSuccess(result)
+        }
+
+        override fun onError(e: Exception) {
+            activityRef.get()?.onConfirmError(e)
+        }
+    }
+}
