+++ example/src/main/java/com/stripe/example/activity/SimplePaymentMethodConfirmationActivity.kt
@@ -1,8 +1,6 @@
 package com.stripe.example.activity
 
-import android.app.Application
 import android.content.Context
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -10,44 +8,17 @@ import android.view.ViewGroup
 import android.widget.ArrayAdapter
 import android.widget.Filter
 import androidx.annotation.DrawableRes
-import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.AndroidViewModel
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
-import com.stripe.android.ApiResultCallback
-import com.stripe.android.PaymentConfiguration
-import com.stripe.android.PaymentIntentResult
-import com.stripe.android.Stripe
-import com.stripe.android.model.ConfirmPaymentIntentParams
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodCreateParams
 import com.stripe.example.R
 import com.stripe.example.databinding.DropdownMenuPopupItemBinding
 import com.stripe.example.databinding.SimplePaymentMethodActivityBinding
-import com.stripe.example.module.BackendApiFactory
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.disposables.CompositeDisposable
-import io.reactivex.schedulers.Schedulers
-import org.json.JSONObject
-import retrofit2.HttpException
-import java.lang.ref.WeakReference
 
-class SimplePaymentMethodConfirmationActivity : AppCompatActivity() {
+class SimplePaymentMethodConfirmationActivity : PaymentIntentActivity() {
 
-    private val stripe: Stripe by lazy {
-        Stripe(this, PaymentConfiguration.getInstance(this).publishableKey)
-    }
     private val viewBinding: SimplePaymentMethodActivityBinding by lazy {
         SimplePaymentMethodActivityBinding.inflate(layoutInflater)
-    }
-    private val keyboardController: KeyboardController by lazy {
-        KeyboardController(this)
-    }
-    private val viewModel: SimplePaymentMethodConfirmationViewModel by lazy {
-        ViewModelProvider(this,
-            ViewModelProvider.AndroidViewModelFactory(application)
-        )[SimplePaymentMethodConfirmationViewModel::class.java]
     }
 
     private val dropdownItem: DropdownItem
@@ -100,11 +71,7 @@ class SimplePaymentMethodConfirmationActivity : AppCompatActivity() {
         onDropdownItemSelected()
 
         viewBinding.payNow.setOnClickListener {
-            keyboardController.hide()
-
-            viewModel.createPaymentIntent(dropdownItem.country) {
-                handleCreatePaymentIntentResponse(it, paymentMethodCreateParams)
-            }
+            createAndConfirmPaymentIntent(dropdownItem.country, paymentMethodCreateParams)
         }
     }
 
@@ -125,45 +92,11 @@ class SimplePaymentMethodConfirmationActivity : AppCompatActivity() {
         }
     }
 
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        stripe.onPaymentResult(requestCode, data, PaymentIntentResultCallback(this))
-    }
-
     private fun enableUi(enabled: Boolean) {
         viewBinding.payNow.isEnabled = enabled
         viewBinding.name.isEnabled = enabled
         viewBinding.email.isEnabled = enabled
         viewBinding.progressBar.visibility = viewVisibility(!enabled)
-    }
-
-    private fun handleCreatePaymentIntentResponse(
-        responseData: JSONObject,
-        params: PaymentMethodCreateParams
-    ) {
-        val secret = responseData.getString("secret")
-        viewModel.status.value += "\n\nStarting PaymentIntent confirmation"
-        stripe.confirmPayment(
-            this,
-            ConfirmPaymentIntentParams.createWithPaymentMethodCreateParams(
-                paymentMethodCreateParams = params,
-                clientSecret = secret,
-                returnUrl = "example://return_url"
-            )
-        )
-    }
-
-    private fun onConfirmSuccess(result: PaymentIntentResult) {
-        val paymentIntent = result.intent
-        viewModel.status.value += "\n\n" +
-            "PaymentIntent confirmation outcome: ${result.outcome}\n\n" +
-            getString(R.string.payment_intent_status, paymentIntent.status)
-        viewModel.inProgress.value = false
-    }
-
-    private fun onConfirmError(e: Exception) {
-        viewModel.status.value += "\n\nException: " + e.message
-        viewModel.inProgress.value = false
     }
 
     companion object {
@@ -188,21 +121,6 @@ class SimplePaymentMethodConfirmationActivity : AppCompatActivity() {
                 PaymentMethodCreateParams.Companion::createEps),
             Giropay("de", R.drawable.ic_brandicon__giropay,
                 PaymentMethodCreateParams.Companion::createGiropay);
-        }
-
-        private class PaymentIntentResultCallback(
-            activity: SimplePaymentMethodConfirmationActivity
-        ) : ApiResultCallback<PaymentIntentResult> {
-
-            private val activityRef = WeakReference(activity)
-
-            override fun onSuccess(result: PaymentIntentResult) {
-                activityRef.get()?.onConfirmSuccess(result)
-            }
-
-            override fun onError(e: Exception) {
-                activityRef.get()?.onConfirmError(e)
-            }
         }
 
         private class DropdownItemAdapter(
@@ -249,55 +167,6 @@ class SimplePaymentMethodConfirmationActivity : AppCompatActivity() {
 
                 override fun publishResults(p0: CharSequence?, p1: FilterResults?) {
                 }
-            }
-        }
-
-        internal class SimplePaymentMethodConfirmationViewModel(
-            application: Application
-        ) : AndroidViewModel(application) {
-            val inProgress = MutableLiveData<Boolean>()
-            val status = MutableLiveData<String>()
-
-            private val context = application.applicationContext
-            private val backendApi = BackendApiFactory(context).create()
-            private val compositeSubscription = CompositeDisposable()
-
-            fun createPaymentIntent(country: String, callback: (JSONObject) -> Unit) {
-                compositeSubscription.add(
-                    backendApi.createPaymentIntent(
-                        mutableMapOf(
-                            "country" to country
-                        )
-                    )
-                        .doOnSubscribe {
-                            inProgress.postValue(true)
-                            status.postValue(context.getString(R.string.creating_payment_intent))
-                        }
-                        .map {
-                            JSONObject(it.string())
-                        }
-                        .subscribeOn(Schedulers.io())
-                        .observeOn(AndroidSchedulers.mainThread())
-                        .subscribe(
-                            {
-                                status.postValue(status.value + "\n\n" +
-                                    context.getString(R.string.payment_intent_status,
-                                        it.getString("status")))
-                                callback(it)
-                            },
-                            {
-                                val errorMessage =
-                                    (it as? HttpException)?.response()?.errorBody()?.string()
-                                        ?: it.message
-                                status.postValue(status.value + "\n\n$errorMessage")
-                                inProgress.postValue(false)
-                            }
-                        ))
-            }
-
-            override fun onCleared() {
-                super.onCleared()
-                compositeSubscription.dispose()
             }
         }
     }
